+++ fe/fe-core/src/main/java/org/apache/doris/common/MetaReader.java
@@ -79,7 +79,7 @@ public class MetaReader {
             checksum = Catalog.getCurrentSystemInfo().loadBackends(dis, checksum);
             checksum = catalog.loadDb(dis, checksum);
             // ATTN: this should be done after load Db, and before loadAlterJob
-            catalog.recreateTabletInvertIndex();
+            catalog.getInternalDataSource().recreateTabletInvertIndex();
             // rebuild es state state
             catalog.getEsRepository().loadTableFromCatalog();
             checksum = catalog.loadLoadJob(dis, checksum);
